Move error helper into the common error folder
