+++ src/server/exemption/project-name/utils.js
@@ -1,15 +1,4 @@
 /**
- * Outputs all errors to an object with the field name as the key
- * Used in the view template to show errors inline
- */
-export const errorDescriptionByFieldName = (errors = []) => {
-  return errors.reduce((error, obj) => {
-    error[obj.field] = obj
-    return error
-  }, {})
-}
-
-/**
  * Converts back end error codes to a user readable description
  */
 export const mapErrorMessage = (error) => {
