+++ src/server/exemption/project-name/utils.test.js
@@ -1,44 +1,6 @@
-import {
-  errorDescriptionByFieldName,
-  mapErrorMessage
-} from '~/src/server/exemption/project-name/utils.js'
+import { mapErrorMessage } from '~/src/server/exemption/project-name/utils.js'
 
 describe('#projectNameController/utils', () => {
-  describe('#projectNameController/utils/errorDescriptionByFieldName', () => {
-    test('return error formatted correctly for front end display', () => {
-      const result = errorDescriptionByFieldName()
-      expect(result).toEqual({})
-    })
-
-    test('return empty array when no data is provided', () => {
-      const result = errorDescriptionByFieldName([
-        {
-          href: '#test',
-          text: 'Test text',
-          field: 'test'
-        },
-        {
-          href: '#test2',
-          text: 'Test text 2',
-          field: 'test2'
-        }
-      ])
-
-      expect(result).toEqual({
-        test: {
-          href: '#test',
-          text: 'Test text',
-          field: 'test'
-        },
-        test2: {
-          href: '#test2',
-          text: 'Test text 2',
-          field: 'test2'
-        }
-      })
-    })
-  })
-
   describe('#projectNameController/utils/mapErrorMessage', () => {
     test('return input value if no error is specified', () => {
       const result = mapErrorMessage('Error')
